backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
mysqlclient==2.2.4
python-decouple==3.8
django-cors-headers==4.7.0
requests==2.31.0